+++ s3.tf
@@ -48,9 +48,9 @@ resource "aws_s3_bucket" "apiary_data_bucket" {
     }
 
     dynamic "expiration" {
-    for_each = lookup(var.apiary_managed_schemas[count.index], "s3_object_expiration_days", var.s3_object_expiration_days) != null ? [1] : []
+    for_each = lookup(var.apiary_managed_schemas[count.index], "s3_object_expiration_days", null) != null ? [1] : []
     content {
-      days = lookup(var.apiary_managed_schemas[count.index], "s3_object_expiration_days", var.s3_object_expiration_days)
+      days = lookup(var.apiary_managed_schemas[count.index], "s3_object_expiration_days", null)
       }
     }
   }